+++ modules/self-hosted/bookstack.nix
@@ -1,7 +1,7 @@
 { config, lib, pkgs, ... }:
 
 let
-  bookstack-secrets = config.sops.secrets."bookstack-secrets".path;
+  bookstack-secrets = config.ghostship.selfHostedSecrets.units."bookstack-secrets".path;
 in
 {
   virtualisation.oci-containers.containers."bookstack" = {